widont: Pass failing tests for short texts
This perhaps isn't as good of a solution as it could be since it doesn't tie
into the existing regexp, but it is simple and doesn't trample on any of the
other tests.  It also allows me to provide awareness of the problem and provide
a proposed solution, without investing too much time in trying to modify the
regexp.